src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-whitespace-attach-expected.txt

CONSOLE ERROR: line 5: Uncaught TypeError: host.createShadowRoot is not a function
